Django_server/requirements.txt

beautifulsoup4==4.10.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
da-vinci==0.3.0
Django==3.2.8
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-environ==0.8.1
django-filter==21.1
django-phonenumber-field==5.2.0
django-thumbnails==0.6.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
flake8==4.0.1
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
isort==5.9.3
lazy-object-proxy==1.6.0
lxml==4.6.3
Markdown==3.3.4
mccabe==0.6.1
multidict==5.2.0
packaging==21.0
phonenumbers==8.12.35
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.10.0
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
python-gettext==4.0
pytz==2021.3
regex==2021.10.8
requests==2.26.0
rsa==4.7.2
shortuuid==1.0.1
sorl-thumbnail==12.7.0
soupsieve==2.2.1
sqlparse==0.4.2
telepot==12.7
Telethon==1.23.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
wrapt==1.13.2
yarl==1.7.0
sentry-sdk==1.4.3
django-modeltranslation==0.17.3
python-memcached==1.59
